feat: Add env differentation to test result
This feature is being done in order to support differentiating the
same test being run in a different environment. This is done by adding
the env field to the Test data model, and setting that env field to
be a hash of the flags and job code of the upload a TestInstance is
found in. The uniqueness constraint of Test now includes the `env` field.

Signed-off-by: joseph-sentry <joseph.sawaya@sentry.io>