fix: save all testruns but only comment relevant ones
This commit adds the `active` field to the TestInstance model.

This field is used to track whether a TestInstance is the "latest"
one for its `Test` in a given commit.

When a TestInstance is created, it is active by default.
If there already exists a TestInstance that be to the same test,
their run_numbers are compared to determine which one is more recent.
The less recent is made inactive.

The notify task only queries for active failures to decide whether
it should comment or not.

Signed-off-by: joseph-sentry <joseph.sawaya@sentry.io>