perf: move test result db writes back to processor
This commit moves the writes in the test result finisher back to the
test result processor task.

The first step is to change the primary key of the Test object to a
text field. This is so we can use a computed primary key for Test
objects.

By computing the test ID using the generate_test_id function,
we avoid having to use a returning clause from the insert on conflict
do nothing for Test objects.

The timestamp field on the TestInstance object has also been removed
as its replaced by using the upload created_at to determine the order
of TestInstance objects to be showed in the PR comment.

Signed-off-by: joseph-sentry <joseph.sawaya@sentry.io>